Fix an actual circular import because of wildcard import
